cleaned tu delft repo and added readme
